# Conflicts:
#	globalshop-biz1-app/src/main/java/com/wangqin/globalshop/biz1/app/dal/dataObject/ShippingTrackDO.java
#	globalshop-biz1-app/src/main/java/com/wangqin/globalshop/biz1/app/dal/dataObject/WarehouseDO.java
#	globalshop-biz1-app/src/main/java/com/wangqin/globalshop/biz1/app/dal/mapperExt/ChannelAccountDOMapperExt.java
#	globalshop-biz1-app/src/main/java/com/wangqin/globalshop/biz1/app/dal/mapperExt/ItemSkuMapperExt.java
#	globalshop-biz1-app/src/main/java/com/wangqin/globalshop/biz1/app/dal/mapperExt/MallOrderMapperExt.java
#	globalshop-biz1-app/src/main/resources/dal/mapper/ShippingTrackDOMapper.xml
#	globalshop-biz1-app/src/main/resources/dal/mapper/WarehouseDOMapper.xml
#	globalshop-biz1-app/src/main/resources/dal/mapperExt/ChannelMallSubOrderDOMapperExt.xml
#	globalshop-biz1-app/src/main/resources/dal/mapperExt/ItemSkuMapperExt.xml
#	globalshop-channel/pom.xml
#	globalshop-channel/src/main/java/com/wangqin/globalshop/channel/dal/mapperExt/CAChannelAccountDOMapperExt.java
#	globalshop-channel/src/main/java/com/wangqin/globalshop/channel/service/channel/IChannelService.java
#	globalshop-channel/src/main/java/com/wangqin/globalshop/channel/service/channelAccount/ChannelAccountServiceImpl.java
#	globalshop-item-app/resources/config/app.properties
#	globalshop-item-app/resources/dal/mapper/AuthOrganizationDOMapper.xml
#	globalshop-item-app/resources/dal/mapper/AuthResourceDOMapper.xml
#	globalshop-item-app/resources/dal/mapper/AuthRoleDOMapper.xml
#	globalshop-item-app/resources/dal/mapper/AuthRoleResourceDOMapper.xml
#	globalshop-item-app/resources/dal/mapper/AuthUserDOMapper.xml
#	globalshop-item-app/resources/dal/mapper/AuthUserRoleDOMapper.xml
#	globalshop-item-app/resources/dal/mapper/BuyerDOMapper.xml
#	globalshop-item-app/resources/dal/mapper/BuyerEntryManifestDOMapper.xml
#	globalshop-item-app/resources/dal/mapper/BuyerReceiptDOMapper.xml
#	globalshop-item-app/resources/dal/mapper/BuyerReceiptDetailDOMapper.xml
#	globalshop-item-app/resources/dal/mapper/BuyerStorageDOMapper.xml
#	globalshop-item-app/resources/dal/mapper/BuyerStorageDetailDOMapper.xml
#	globalshop-item-app/resources/dal/mapper/BuyerTaskDOMapper.xml
#	globalshop-item-app/resources/dal/mapper/BuyerTaskDetailDOMapper.xml
#	globalshop-item-app/resources/dal/mapper/ChannelAccountConfigDOMapper.xml
#	globalshop-item-app/resources/dal/mapper/ChannelAccountDOMapper.xml
#	globalshop-item-app/resources/dal/mapper/ChannelDOMapper.xml
#	globalshop-item-app/resources/dal/mapper/ChannelListingItemDOMapper.xml
#	globalshop-item-app/resources/dal/mapper/ChannelListingItemSkuDOMapper.xml
#	globalshop-item-app/resources/dal/mapper/ChannelSalePriceDOMapper.xml
#	globalshop-item-app/resources/dal/mapper/ChannelSessionDOMapper.xml
#	globalshop-item-app/resources/dal/mapper/CompanyDOMapper.xml
#	globalshop-item-app/resources/dal/mapper/CountryDOMapper.xml
#	globalshop-item-app/resources/dal/mapper/DealerDOMapper.xml
#	globalshop-item-app/resources/dal/mapper/DealerTypeDOMapper.xml
#	globalshop-item-app/resources/dal/mapper/InventoryBookingRecordDOMapper.xml
#	globalshop-item-app/resources/dal/mapper/InventoryDOMapper.xml
#	globalshop-item-app/resources/dal/mapper/InventoryInoutDOMapper.xml
#	globalshop-item-app/resources/dal/mapper/InventoryOnWareHouseDOMapper.xml
#	globalshop-item-app/resources/dal/mapper/InventoryOutManifestDOMapper.xml
#	globalshop-item-app/resources/dal/mapper/InventoryOutManifestDetailDOMapper.xml
#	globalshop-item-app/resources/dal/mapper/ItemBrandDOMapper.xml
#	globalshop-item-app/resources/dal/mapper/ItemCategoryDOMapper.xml
#	globalshop-item-app/resources/dal/mapper/ItemCategoryScaleDOMapper.xml
#	globalshop-item-app/resources/dal/mapper/ItemCommentDOMapper.xml
#	globalshop-item-app/resources/dal/mapper/ItemDOMapper.xml
#	globalshop-item-app/resources/dal/mapper/ItemFindDOMapper.xml
#	globalshop-item-app/resources/dal/mapper/ItemSkuDOMapper.xml
#	globalshop-item-app/resources/dal/mapper/ItemSkuScaleDOMapper.xml
#	globalshop-item-app/resources/dal/mapper/LogisticCategoryMappingDOMapper.xml
#	globalshop-item-app/resources/dal/mapper/LogisticCompanyDOMapper.xml
#	globalshop-item-app/resources/dal/mapper/MallCustomerDOMapper.xml
#	globalshop-item-app/resources/dal/mapper/MallOrderDOMapper.xml
#	globalshop-item-app/resources/dal/mapper/MallReturnOrderDOMapper.xml
#	globalshop-item-app/resources/dal/mapper/MallShippingAddressDOMapper.xml
#	globalshop-item-app/resources/dal/mapper/MallShippingCartDOMapper.xml
#	globalshop-item-app/resources/dal/mapper/MallWxBannerDOMapper.xml
#	globalshop-item-app/resources/dal/mapper/MallWxCustomerTrackDOMapper.xml
#	globalshop-item-app/resources/dal/mapper/MallWxItemSnapshotDOMapper.xml
#	globalshop-item-app/resources/dal/mapper/MallWxPayBillDOMapper.xml
#	globalshop-item-app/resources/dal/mapper/MallWxSpecialPageDOMapper.xml
#	globalshop-item-app/resources/dal/mapper/ReactVersionDOMapper.xml
#	globalshop-item-app/resources/dal/mapper/SequenceDOMapper.xml
#	globalshop-item-app/resources/dal/mapper/ShippingOrderDOMapper.xml
#	globalshop-item-app/resources/dal/mapper/ShippingPackingPatternDOMapper.xml
#	globalshop-item-app/resources/dal/mapper/ShippingPackingScaleDOMapper.xml
#	globalshop-item-app/resources/dal/mapper/ShippingTrackDOMapper.xml
#	globalshop-item-app/resources/dal/mapper/ShippingTrackPolling4pxDOMapper.xml
#	globalshop-item-app/resources/dal/mapper/ShippingTrackPollingYuntongDOMapper.xml
#	globalshop-item-app/resources/dal/mapperExt/BuyerStorageDetailMapperExt.xml
#	globalshop-item-app/resources/dal/mapperExt/BuyerTaskDetailMapperExt.xml
#	globalshop-item-app/resources/dal/mapperExt/ChannelAccountConfigDOMapperExt.xml
#	globalshop-item-app/resources/dal/mapperExt/ChannelAccountDOMapperExt.xml
#	globalshop-item-app/resources/dal/mapperExt/CompanyDOMapperExt.xml
#	globalshop-item-app/resources/dal/mapperExt/InventoryMapperExt.xml
#	globalshop-item-app/resources/dal/mapperExt/InventoryOnWarehouseMapperExt.xml
#	globalshop-item-app/resources/dal/mapperExt/ItemBrandDOMapperExt.xml
#	globalshop-item-app/resources/dal/mapperExt/ItemDOMapperExt.xml
#	globalshop-item-app/resources/dal/mapperExt/ItemMapperExt.xml
#	globalshop-item-app/resources/dal/mapperExt/ItemSkuMapperExt.xml
#	globalshop-item-app/resources/dal/mapperExt/MallWxBannerDOMapperExt.xml
#	globalshop-item-app/resources/dal/mapperExt/MallWxSpecialPageDOMapperExt.xml
#	globalshop-item-app/resources/dal/mapperExt/SequenceMapperExt.xml
#	globalshop-item-app/resources/dal/mapperExt/ShippingPackingPatternExtMapper.xml
#	globalshop-item-app/resources/dal/mapperExt/mallCustomerDOMapperExt.xml
#	globalshop-item-app/resources/dal/spring-dal.xml
#	globalshop-item-app/resources/generatorConfig.xml
#	globalshop-item-app/resources/spring/spring-app.xml
#	globalshop-item-app/resources/spring/spring-dubbo-consumer.xml
#	globalshop-item-app/resources/spring/spring-dubbo-provider.xml
#	globalshop-item-app/src/main/java/com/wangqin/globalshop/item/app/service/IInventoryService.java
#	globalshop-item-app/src/main/java/com/wangqin/globalshop/item/app/service/impl/InventoryServiceImpl.java
#	globalshop-order-app/src/main/java/com/wangqin/globalshop/order/app/controller/ShippingOrderController.java
#	globalshop-usercenter-app/pom.xml

+++ globalshop-order-app/src/main/java/com/wangqin/globalshop/order/app/service/shipping/impl/ShippingOrderServiceImpl.java
@@ -310,13 +310,13 @@ public class ShippingOrderServiceImpl implements IShippingOrderService {
                         }
                     }
                     if(closeCount == totalCount) {	//主订单关闭
-                        outerOrder.setStatus((byte) OrderStatus.CLOSE.getCode());
+                        outerOrder.setStatus(OrderStatus.CLOSE.getCode());
                     } else if((closeCount+sentCount) == totalCount) {	//全部发货
-                        outerOrder.setStatus((byte) OrderStatus.SENT.getCode());
+                        outerOrder.setStatus(OrderStatus.SENT.getCode());
                     } else if(sentCount > 0) {	//部分发货
-                        outerOrder.setStatus((byte) OrderStatus.PART_SENT.getCode());
+                        outerOrder.setStatus( OrderStatus.PART_SENT.getCode());
                     } else {
-                        outerOrder.setStatus((byte) OrderStatus.INIT.getCode());
+                        outerOrder.setStatus( OrderStatus.INIT.getCode());
                     }
                     outerOrder.setGmtModify(new Date());
                     mallOrderService.updateById(outerOrder);
